auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.4.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0